API-Agente/requirements.txt
# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Gemini/Google AI
google-generativeai>=0.3.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0